# ci: faster checkers build