@dbolduc 2023-01-05
> the extra ~14s

14s is longer than it takes to type `env GOOGLE_CLOUD_CPP_FORMAT_CHANGES_ONLY=1` :shrug:

also our repo will only grow in size, so the build will only get longer. and there is opportunity to conditionally run the typos / markdown generator / doxygen update steps.

> You wouldn't do an actual build there, right? So why do a `checkers` there?

I make code changes on my laptop, then send them off to the cloudtop in order to run a build. I do not run `checkers` on the cloudtop because I want to keep the changes made by `checkers` locally. (I am not saying this workflow is smart. I am just saying it is what I do).