@devbww 2023-01-05
> I make code changes on my laptop, then send them off to the cloudtop in order to run a build. I do not run `checkers` on the cloudtop because I want to keep the changes made by `checkers` locally. (I am not saying this workflow is smart. I am just saying it is what I do).

OK ... fair enough.