@dbolduc 2023-01-05
> "there is opportunity to conditionally run the typos / markdown generator / doxygen update steps."

I will try to conditionally skip the `ci/generate-markdown/...` business in a follow up PR, tomorrow.